@@ -1,17 +1,20 @@
 name: http_multi_server
-version: 2.2.1-dev
+version: 3.0.0-dev
 
 description: >-
   A dart:io HttpServer wrapper that handles requests from multiple servers.
 homepage: https://github.com/dart-lang/http_multi_server
 
 environment:
-  sdk: '>=2.1.1 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
-  async: '>=1.2.0 <3.0.0'
+  async: ^2.5.0
 
 dev_dependencies:
-  http: ^0.12.0
-  pedantic: ^1.8.0
-  test: ^1.5.2
+  http: ^0.13.0
+  pedantic: ^1.10.0
+  test: ^1.16.0
+
+dependency_overrides:
+  test: ^1.16.0
